+++ src/flat_functor_tmpl.hpp
@@ -35,16 +35,16 @@ TMPL_EXTERN template size_t flat_sparse_tensor_matmul<TMPL_Q, TMPL_FL>(
 #else
 
 TMPL_EXTERN template tuple<py::array_t<uint32_t>, py::array_t<uint32_t>,
-                           py::array_t<double>, py::array_t<uint32_t>>
+                           py::array_t<double>, py::array_t<uint64_t>>
 flat_sparse_tensor_diag<TMPL_Q>(const py::array_t<uint32_t> &aqs,
                                 const py::array_t<uint32_t> &ashs,
                                 const py::array_t<double> &adata,
-                                const py::array_t<uint32_t> &aidxs,
+                                const py::array_t<uint64_t> &aidxs,
                                 const py::array_t<int> &idxa,
                                 const py::array_t<int> &idxb);
 
 TMPL_EXTERN template tuple<py::array_t<uint32_t>, py::array_t<uint32_t>,
-                           py::array_t<uint32_t>>
+                           py::array_t<uint64_t>>
 flat_sparse_tensor_tensordot_skeleton<TMPL_Q>(const py::array_t<uint32_t> &aqs,
                                               const py::array_t<uint32_t> &ashs,
                                               const py::array_t<uint32_t> &bqs,
@@ -63,10 +63,10 @@ flat_sparse_tensor_matmul_init<TMPL_Q>(
 TMPL_EXTERN template py::array_t<int32_t>
 flat_sparse_tensor_matmul_plan<TMPL_Q>(
     const py::array_t<uint32_t> &aqs, const py::array_t<uint32_t> &ashs,
-    const py::array_t<uint32_t> &aidxs, const py::array_t<uint32_t> &bqs,
-    const py::array_t<uint32_t> &bshs, const py::array_t<uint32_t> &bidxs,
+    const py::array_t<uint64_t> &aidxs, const py::array_t<uint32_t> &bqs,
+    const py::array_t<uint32_t> &bshs, const py::array_t<uint64_t> &bidxs,
     const py::array_t<int> &idxa, const py::array_t<int> &idxb,
-    const py::array_t<uint32_t> &cqs, const py::array_t<uint32_t> &cidxs,
+    const py::array_t<uint32_t> &cqs, const py::array_t<uint64_t> &cidxs,
     bool ferm_op);
 
 #endif
